Extract alias logic to separate file
